Move `DeleteResponses` command
