+++ wcfsetup/install/files/lib/system/moderation/queue/AbstractCommentResponseModerationQueueHandler.class.php
@@ -158,7 +158,7 @@ class AbstractCommentResponseModerationQueueHandler extends AbstractCommentComme
     public function removeContent(ModerationQueue $queue, $message)
     {
         if ($this->isValid($queue->objectID)) {
-            (new \wcf\system\comment\response\command\DeleteResponses([$this->getResponse($queue->objectID)]))();
+            (new \wcf\command\comment\response\DeleteResponses([$this->getResponse($queue->objectID)]))();
         }
     }
